Safari-609.1.11.1/LayoutTests/http/tests/websocket/tests/hybi/workers/multiple-subprotocols.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../../js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("Test WebSocket subprotocol negotiation in workers.");

window.jsTestIsAsync = true;

function startsWith(str, prefix)
{
    return str.indexOf(prefix) == 0;
}

var worker = new Worker("resources/multiple-subprotocols.js");
worker.onmessage = function (event)
{
    var message = event.data;
    if (startsWith(message, "PASS"))
        testPassed(message);
    else if (startsWith(message, "FAIL"))
        testFailed(message)
    else
        debug(message);
    if (message === "DONE")
        finishJSTest();
};

</script>
<script src="../../../../../js-test-resources/js-test-post.js"></script>
</body>
</html>
