LayoutTests/http/tests/websocket/tests/hybi/workers/multiple-subprotocols-expected.txt

Test WebSocket subprotocol negotiation in workers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

INFO: Connected
INFO: Received: superchat
PASS PASS: receivedMessage is equal to "superchat"
INFO: Closed
PASS PASS: closeEvent.wasClean is true
DONE
PASS successfullyParsed is true

TEST COMPLETE
